renamed searchBlogsList to blogsList
