+++ app/app-module.js
@@ -17,7 +17,7 @@ var app = angular.module('beamer', [
     beamer.common.logoutModal.module.name,
     beamer.common.editRouteModal.module.name,
     beamer.common.confirmDeleteModal.module.name,
-    beamer.common.searchBlogsList.module.name,
+    beamer.common.blogsList.module.name,
     beamer.common.models.module.name,
     beamer.views.profile.module.name,
     beamer.views.blog.module.name,
