+++ app/common/blogsList/blogs-list-directive.js
@@ -1,0 +1,23 @@
+var BlogsListCtrl = require('./blogs-list-controller');
+
+var blogsList = function() {
+
+    return {
+        restrict: 'E',
+        scope: {
+            blogs: '=',
+            noBlogs: '&'
+        },
+        templateUrl: 'common/blogsList/blogs-list-directive.html',
+        controller: BlogsListCtrl,
+        controllerAs: 'blogsListCtrl',
+        link: function(scope, e, a, controller) {
+            scope.$watch('blogs', function(newVal) {
+                controller.blogs = newVal;
+            })
+        }
+    }
+
+}
+
+module.exports = blogsList;
