+++ app/common/blogsList/blogs-list.js
@@ -1,0 +1,4 @@
+var blogsList = require('./blogs-list-directive.js');
+
+module.exports = angular.module('beamer.common.blogsList', [])
+    .directive('blogsList', [blogsList]);
