--- app/common/searchBlogsList/search-blogs-list.js
@@ -1,4 +1,0 @@
-var searchBlogsList = require('./search-blogs-list-directive.js');
-
-module.exports = angular.module('beamer.common.searchBlogsList', [])
-    .directive('searchBlogsList', [searchBlogsList]);
